refactor: 상품 상세 페이지 컴포넌트 파일 분할
