+++ src/components/ItemDetail/ItemDetail.module.scss
@@ -1,0 +1,207 @@
+.Item {
+  display: flex;
+  gap: 40px;
+  border-bottom: 1px solid var(--gray-300);
+  padding-bottom: 40px;
+}
+
+.main {
+  flex-grow: 1;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.image {
+  width: 486px;
+  height: 486px;
+  border-radius: 16px;
+}
+
+.header {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 16px 0;
+
+  border-bottom: 1px solid var(--gray-200);
+  color: var(--gray-800);
+  margin-bottom: 24px;
+
+  .title {
+    flex-grow: 1;
+    font-size: 24px;
+    font-weight: 600;
+    align-content: center;
+  }
+
+  .price {
+    flex: 1 0 100%;
+    font-size: 40px;
+    font-weight: 600;
+    margin-bottom: 16px;
+  }
+}
+
+.description {
+  color: var(--gray-600);
+  margin-bottom: 24px;
+  flex-grow: 1;
+
+  .title {
+    font-size: 16px;
+    font-weight: 600;
+    margin-bottom: 16px;
+  }
+
+  .content {
+    font-size: 16px;
+    font-weight: 400;
+    line-height: 26px;
+  }
+}
+
+.tag-list {
+  color: var(--gray-600);
+  margin-bottom: 62px;
+
+  .title {
+    font-size: 16px;
+    font-weight: 600;
+    margin-bottom: 16px;
+  }
+
+  .content {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 8px;
+
+    .tag {
+      background-color: var(--gray-100);
+      padding: 6px 16px;
+      border-radius: 26px;
+      color: var(--gray-800);
+      font-size: 16px;
+      font-weight: 400;
+    }
+  }
+}
+
+.footer {
+  display: flex;
+
+  .owner {
+    flex-grow: 1;
+    display: flex;
+    border-right: 1px solid var(--secondary-200);
+    font-size: 14px;
+
+    .name {
+      color: var(--gray-600);
+      margin-bottom: 2px;
+    }
+
+    .date {
+      color: var(--gray-400);
+    }
+
+    .profile {
+      margin-right: 16px;
+    }
+  }
+
+  .favorite {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 10px;
+
+    padding: 4px 12px;
+    border: 1px solid var(--secondary-200);
+    border-radius: 35px;
+    margin-left: 24px;
+
+    color: var(--gray-500);
+    font-size: 16px;
+    font-weight: 500;
+
+    img {
+      width: 32px;
+    }
+  }
+}
+
+@media screen and (max-width: 1199px) {
+  .container {
+    width: 696px;
+  }
+
+  .Item {
+    gap: 32px;
+    padding-bottom: 32px;
+  }
+
+  .image {
+    width: 340px;
+    height: 340px;
+  }
+
+  .header {
+    gap: 8px 0;
+
+    .title {
+      font-size: 20px;
+    }
+
+    .price {
+      font-size: 32px;
+    }
+  }
+
+  .description {
+    .title {
+      font-size: 14px;
+    }
+  }
+
+  .tag-list {
+    .title {
+      font-size: 14px;
+      margin-bottom: 8px;
+    }
+  }
+
+  .footer {
+    .owner {
+      .profile {
+        img {
+          width: 40px;
+        }
+      }
+    }
+
+    .favorite {
+      gap: 4px;
+
+      img {
+        width: 24px;
+      }
+    }
+  }
+}
+
+@media screen and (max-width: 767px) {
+  .container {
+    width: 344px;
+  }
+
+  .Item {
+    flex-direction: column;
+    gap: 16px;
+    padding-bottom: 24px;
+  }
+
+  .image {
+    width: 343px;
+    height: 343px;
+  }
+}
